requirements.txt
# -*- coding: utf-8 -*-


pytorch-lightning==1.2.10
tensorboard==2.9.0
matplotlib==3.5.2
numpy==1.22.4
librosa==0.9.1
h5py==3.6.0
pandas==1.4.2
tqdm==4.64.0
black==21.12b0
click==8.0.4
pyyaml==6.0
soundfile==0.10.3.post1

advertorch==0.2.3
hydra-core==1.1.2
hydra-colorlog==1.2.0
